pyproject.toml
[tool.poetry]
name = "fastapi-mini"
version = "0.1.0"
description = ""
authors = ["shinbilee <dltlsql8389@naver.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.8"
sqlalchemy = "^2.0.37"
uvicorn = "^0.34.0"
alembic = "^1.14.1"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
databases = "^0.9.0"
pydantic-settings = "^2.7.1"
python-dotenv = "^1.0.1"
pydantic = {extras = ["email"], version = "^2.10.6"}
bcrypt = "^4.2.1"
passlib = "^1.7.4"
httpx = "^0.28.1"
pytest-asyncio = "^0.25.3"
aiosqlite = "^0.21.0"
greenlet = "^3.1.1"
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pydantic = "^2.10.6"
isort = "^6.0.0"
ruff = "^0.9.4"
pre-commit = "^4.1.0"
httpx = "^0.28.1"
pytest-cov = "^6.0.0"

[tool.ruff]
line-length = 120
exclude = [
    ".git",
    "__pycache__",
    "tests/*"
]
lint.ignore = ["F401", "E402", "E501"]
lint.select = ["I", "E", "F", "W", "C90"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
